backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
openai==1.12.0
anthropic==0.18.1
httpx==0.26.0
aiohttp==3.9.3
reportlab==4.1.0
Pillow==10.2.0
markdown==3.5.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
sse-starlette==1.8.2
python-pptx>=1.0.2
psd-tools>=1.9.0
numpy>=1.26.0
